Begin work toward basic styling of app
