+++ client-src/common/components/footer/footer.styl
@@ -1,0 +1,12 @@
+footer
+  position fixed
+  bottom 0
+  left 0
+  width 100%
+  background-color #fff
+  border-bottom 1px solid #dedede
+
+  > .container
+    height 49px
+    padding 0 8px
+    line-height 49px
