+++ client-src/common/components/header/header.styl
@@ -1,12 +1,17 @@
 .app-header
-  margin 20px 0
-  padding-bottom 20px
+  position fixed
+  width 100%
+  background-color #fff
   border-bottom 1px solid #dedede
 
-  .app-logo
-    margin 0
-    padding 0
+  > .container
+    height 40px
+    padding 0 8px
+    display flex
+    justify-content flex-end
 
-  .app-logo-link
-    text-decoration none
-    color #000
+  .create-new-transaction
+    color #4cc1fc
+    border 0
+    background #fff
+    font-size 16px
