+++ client-src/common/stylus/scaffolding.styl
@@ -1,15 +1,17 @@
 html, body
   margin 0
   padding 0
-
-body
-  min-width $appWidth + $appPadding
+  background-color #f5f5f5
 
 .container
-  width $appWidth
+  max-width $appWidth
   margin 0 auto
 
 main
+  // 40px from the header + 20px margin
+  padding-top 60px
   margin-bottom 20px
   padding-bottom 20px
   border-bottom 1px solid #dedede
+  // Temporary code to test scrolling behavior on mobile
+  min-height 750px
