+++ client-src/common/stylus/typography.styl
@@ -1,6 +1,7 @@
 html
   font-size $rootFontSize
   color $textColor
+  font-family 'Source Sans Pro', sans-serif
 
 a
   color $linkColor
