+++ client-src/home/components/transactions/index.js
@@ -1,6 +1,5 @@
 import yo from 'yo-yo';
 import store from '../../../redux/store';
-import {createTransaction} from '../../../redux/transactions/action-creators';
 import Transaction from '../transaction';
 
 function getTransactionsList(transactions) {
@@ -9,14 +8,6 @@ function getTransactionsList(transactions) {
 
 function getEmptyTransactions() {
   return yo`<div>Ain't nothin'</div>`;
-}
-
-function onClickNew() {
-  store.dispatch(createTransaction({
-    value: Math.random(),
-    description: 'Hot off the press',
-    date: '2015-10-02'
-  }));
 }
 
 export default function() {
@@ -30,13 +21,8 @@ export default function() {
   }
 
   return yo`
-    <div className="transactions">
-      <div class="transactions-header">
-        <button className="create-transaction" onclick=${onClickNew}>+ New Transaction</button>
-      </div>
-      <ul>
-        ${children}
-      </ul>
-    </div>
+    <ul className="transactions">
+      ${children}
+    </ul>
   `;
 }
