+++ client-src/home/components/transactions/transactions.styl
@@ -1,12 +1,2 @@
 .transactions
-
-  .transactions-header
-    padding-bottom 10px
-    margin-bottom 10px
-
-    .create-transaction
-      border-radius 2px
-      border 1px solid #ccc
-      background #fefefe
-      padding 5px 8px
-      color #888
+  background-color #fff
